Created Line class and refactor Planet class to use Double members
Created a line class with the purpose of getting an easy way to test if
a point is part of a line given by two points.
Refactored the Planet and Galaxy classes to use type Double for
variables xPos, yPos and distance, this made it easier to implement the
Line class.
Added and changed some tests.

+++ src/starTrek/galaxy/Galaxy.java
@@ -6,7 +6,7 @@ package starTrek.galaxy;
 import java.util.ArrayList;
 
 /**
- * @author Molter
+ * @author Mauro Moltrasio
  *
  */
 public class Galaxy {
@@ -48,7 +48,7 @@ public class Galaxy {
 	 * @param speed 	the speed at which the planet rotates
 	 * @param distance 	the distance from the sun to this planet
 	 */
-	public void add(Integer speed, Integer distance) {
+	public void add(Integer speed, Double distance) {
 		this.planets.add(new Planet(speed, distance));
 	}
